feat: Add hooks to logger
