+++ src/core/logger/types.ts
@@ -1,0 +1,10 @@
+import type {LogLevels, LogLevelsAccess} from './logger';
+
+export type LogConsumer = Record<LogLevelsAccess, Writer>;
+
+export type Writer = (...msgs: string[]) => void;
+
+export type LogRecord = {
+    level: LogLevels;
+    message: string;
+};
